+++ src/main/java/org/scion/jpan/ScionDatagramSocket.java
@@ -329,7 +329,7 @@ public class ScionDatagramSocket extends java.net.DatagramSocket {
     synchronized (packet) {
       ByteBuffer receiveBuffer =
           ByteBuffer.wrap(packet.getData(), packet.getOffset(), packet.getLength());
-      ScionSocketAddress responseAddress = channel.receive(receiveBuffer);
+      ScionPathAddress responseAddress = channel.receive(receiveBuffer);
       if (responseAddress == null) {
         // timeout occurred
         throw new SocketTimeoutException();
@@ -404,7 +404,7 @@ public class ScionDatagramSocket extends java.net.DatagramSocket {
     synchronized (packet) {
       ByteBuffer receiveBuffer =
           ByteBuffer.wrap(packet.getData(), packet.getOffset(), packet.getLength());
-      ScionSocketAddress responseAddress = channel.receive(receiveBuffer);
+      ScionPathAddress responseAddress = channel.receive(receiveBuffer);
       if (responseAddress == null) {
         // timeout occurred
         throw new SocketTimeoutException();